move redux hook to store
